Extract complex scripts from SearchService
Moved a lot of redundant scripts into one class.
Judgment processing specific to SearchService, conversion of Object,
refilling processing to list, etc.

The random function is refactored.
The list copy has been changed to a partial copy due to overprocessing.
This will eliminate the dependency on guava.

On the other hand, explicit declarations were added to pom
to use CollectionUtils for addIgnoreNull.