Fixed bug where new users defaulted to access to all media folders
Newly created users were previously given access to all media folders before
the user was updated to only allow access for the folders specified in the
submitted form.
This change now defaults to no access to any folders, in case a crash or
exception occurs between when the user is created and the later update of
the allowed folders.