Extend ffmpegParser to audio files, this enables metadata parsing for less common file formats like APE etc.
This patch enables `FFmpegParser` for all _playable_ file types (defined as all audio and video file types) and enables it to extract metadata from those files. This makes it possible to extract metadata from less common file types like APE which are not supported by `JaudiotaggerParser`. This patch depends on PR #1259 as `FFmpegParser` calls functions which are not present in an unmodified `MetaDataParser`.