Strip terminating and embedded \0 in JaudiotaggerParser
Jaudiotagger sometimes returns strings with trailing and (rarely) embedded \0 characters, strip these before storing them in the database. This avoids problems with Postgresql (which can not store \0 in string data types) and it is more correct as these \0 characters should not be there in the first place. Strangely enough Jaudiotagger *does* strip terminating \0 characters when storing WAV info tags in ID3 fields, this patch extends this functionality to all other (string) types and also takes care of embedded stray \0's.