Simplify the ghetto-semaphore for the cover cache
Synchronising on .intern() is considered a terrible practise and
shouldn't be done. For details, see https://shipilev.net/jvm/anatomy-quarks/10-string-intern/
and https://stackoverflow.com/questions/133988/synchronizing-on-string-objects-in-java

An other possibility would be to use Striped from Guava,
but since the classes in still in beta, I didn't want to use it.

This commit simply extracts the put-image-in-cache part
in a new method, and mark it as synchronised: most of the time,
the image is already in the cache anyway, and I rather have simple code
behaving correctly that we can easily optimize later than hackish crap
that will result in horrible corruptions that nobody will be able
to debug.