Don't run the scanning loop twice
There is no point in scanning once for folders, then for files.
It's faster to scan once, both for folders and files,
because calling getChildrenOf without using the cache will
actually update the cache twice. It's ok from a functional point of view since
the update is idempotent, but since it's accessing the disk to do so,
it significantly hinders performances.