Removed the auto-refresh feature
This feature just generated a lot of unnecessary traffic to the server and
hurt usability because the page would refresh unexpectedly while a user was
actively looking at content.  Fixes #1003, #1179, #729.
Did not remove the column for that setting in the user database, since that
would lead to complications if needing to downgrade for some reason.