Fix erroneous Liquibase migrations
addPrimaryKey with composite keys is for some reason generating SQL (on HSQL) that borrows the name of the fk constraint of one of the columns (chosen arbitrarily). That results in an error because two constraints with the same name conflict.

To fix, just let liquibase know that each column is a primary column, it'll generate the correct SQL. These were the only instances of this pattern found.

To reproduce:
- Start airsonic with no existing db (to run migrations)
- Either examine the SQL produced in the airsonic.script db file, or try restarting airsonic and watch it fail
  - A symptom should exist on line 7 (user_role table)

https://github.com/liquibase/liquibase/pull/167 is a registered bug that says something similar was fixed, but I still see issues