Allow previous Liquibase migration checksums for MariaDB
If someone was using MariaDB before #1188, they were probably doing so
by using the MySQL JDBC driver to pass the migration (which had the
proper workarounds in place for some column types).

This commit assumes that the migrations done that way were valid and
produced the same result as running the MariaDB migrations after #1188.